Clear store state on user logout
Each reducer handles its own reset on LOGOUT_SUCCESS, as resetting the
whole store causes problems for other components which store data there,
such as react-router-redux.

This makes pubsweet-component-login a full dependency of
pubsweet-client, but given that it's already a dependency of the tests I
think that's ok. The alternative would be to just use 'LOGOUT_SUCCESS'
as a string, instead of the constant.