feat(server): remove user.teams pre-filling
When getting the current user from the authentication endpoint, user.teams was prefilled with team
objects. This caused inconsistencies between client and server, that became apparent when a
cross-platform authorization mode was written. This commit removes that pre-filling and so removes
the inconsistency.

BREAKING CHANGE: User.teams no longer returns an array of objects, but instead returns an array of
ids, as it's stored in the database.