feat(sse): introduce authorization-based filtering of server-sent events
BREAKING CHANGE: The pubsweet-sse package is now strongly bound to the authorization library
(Authsome) that PubSweet uses in the server. Authsome is now used to filter events based on
permissions, and by default, events will be dropped/not sent. They will be sent only if handled by
the Authsome mode.