feat(server): add current and update state to update authorization
BREAKING CHANGE: Authsome mode parts that deal with 'PATCH' or update operations, must now deal with
the changed API, which provides a {current:, update:} object, where 'current' represents the current
state of the object, and 'update' representes the requested update. This was added so that one can
verify e.g. that a object.status update is allowed, if the current object.status is 'editing', and
the wished update.status is 'edited'.

fix #393