feat: remove redux
Redux, long deprecated in the client, is removed with this change. It is replaced by a pure Apollo
GraphQL provider.

BREAKING CHANGE: All components that rely on Redux's actions and reducers will cease to work in this
version. Migrating to GraphQL and Apollo on the client and server is advised.