refactor: remove validations from client
BREAKING CHANGE: This removes the concept of validations from the client (shared with server).
Validations in general were removed with the switch to BaseModel, which uses JSON Schema and not Joi
for data validation. While the value of exposing server side validations on the client was never
realized, we might add something similar back in the future for exposing JSON Schema on the
client-side, if a clear use-case is presented.