feat(model-user): move unique constraints verification into db
Adds a migration that adds unique constraints to username and table. Previously these verifications
lived outside of the database (as a result of a PouchDB migration). The User model isn't used by
anyone directly, but is extended in a few apps, an such a migration is considered a breaking change.

BREAKING CHANGE: Moves unique constraints from save()/isUniq() to database-native checks.