feat(job-xsweet): improve API for GraphQL-based conversion
Previously a secondary query was needed after the subscription returned 'Done'. Now the result is
returned/included in the last subscription update (of status 'Done') without the need for a
secondary query.

BREAKING CHANGE: The query for fetching coversion jobs no longer exists, it's purely a mutation +
subscription combo now.

closes #453