feat(model-user): add user to default identity migration
This will create new default identities for all users in the system, carrying over email, username
(as name) and password hashes.

BREAKING CHANGE: This migrates the users table in a destructive way (the data is transferred to the
new table, but no longer exists on the users table). Be careful!