refactor(model-user): drops the framents & collections columns on user
Storing fragment and collection ids is a remnant from the PouchDB system and is no longer used.
Classic foreign key relatioships are to be used instead.

BREAKING CHANGE: If you were depending on the fragments and collections array/column on the User
model itself (vs. custom models with foreign key relationships), then this is a breaking change and
you need to write a migration.