feat(sortable-list): upgrade to latest react-dnd complete
We've upgraded the react-dnd dependency from 2.5.x to 10.0.x and changed the API very slightly for
better ergonomics. Look at the two new live examples in the styleguide for and implementation
reference. Because of these changes this is a breaking change.

BREAKING CHANGE: Some names of props are different (listItem -> ListItem, dragHandle -> DragHandle),
the DndProvider is now used instead of a Dnd context, and you're required to set the ref now (See
examples in
https://gitlab.coko.foundation/pubsweet/pubsweet/tree/master/components/client/component-sortable-list/examples.